+++ packages/expo-observe/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### 🎉 New features
 
+- Add a `react-native-reanimated` integration that reports Reanimated errors and warnings to EAS Observe. ([#50576](https://github.com/expo/expo/pull/50576) by [@douglowder](https://github.com/douglowder))
+
 ### 🐛 Bug fixes
 
 ### 💡 Others